Implement array and append template functions
